pyproject.toml
[project]
name = "pdf2epub-qa"
version = "0.1.0"
description = "PDF to EPUB converter with page-by-page QA review."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "pdf2epub-qa contributors"}]
dependencies = [
    "pymupdf>=1.23.0",
    "ebooklib>=0.18",
    "typer>=0.12.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "pre-commit>=3.7.0",
]
ocr = [
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
pdf2epub = "pdf2epub_qa.cli:app"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
